test(shell): import terminal default directly
