+++ shell/src/stores/terminal-defaults.ts
@@ -1,0 +1,13 @@
+export type TerminalThemeId =
+  | "system"
+  | "one-dark"
+  | "one-light"
+  | "catppuccin-mocha"
+  | "dracula"
+  | "solarized-dark"
+  | "solarized-light"
+  | "nord"
+  | "github-dark"
+  | "github-light";
+
+export const DEFAULT_TERMINAL_THEME_ID: TerminalThemeId = "one-dark";
