+++ shell/src/stores/terminal-settings.ts
@@ -1,22 +1,12 @@
 import { create } from "zustand";
 import { persist } from "zustand/middleware";
+import { DEFAULT_TERMINAL_THEME_ID, type TerminalThemeId } from "./terminal-defaults";
 
-export type TerminalThemeId =
-  | "system"
-  | "one-dark"
-  | "one-light"
-  | "catppuccin-mocha"
-  | "dracula"
-  | "solarized-dark"
-  | "solarized-light"
-  | "nord"
-  | "github-dark"
-  | "github-light";
+export { DEFAULT_TERMINAL_THEME_ID, type TerminalThemeId } from "./terminal-defaults";
 
 export const TERMINAL_FONT_FAMILIES = ["MesloLGS NF", "Berkeley Mono", "JetBrains Mono", "Fira Code"] as const;
 export type TerminalFontFamily = (typeof TERMINAL_FONT_FAMILIES)[number];
 export type TerminalCursorStyle = "block" | "bar" | "underline";
-export const DEFAULT_TERMINAL_THEME_ID: TerminalThemeId = "one-dark";
 
 interface TerminalSettings {
   themeId: TerminalThemeId;
