+++ tests/shell/terminal-settings.test.ts
@@ -1,12 +1,8 @@
 import { describe, expect, it } from "vitest";
-import { readFile } from "node:fs/promises";
+import { DEFAULT_TERMINAL_THEME_ID } from "../../shell/src/stores/terminal-defaults.js";
 
 describe("terminal settings defaults", () => {
-  it("defaults new terminal settings to a dark terminal theme", async () => {
-    const source = await readFile(new URL("../../shell/src/stores/terminal-settings.ts", import.meta.url), "utf8");
-
-    expect(source).toContain('export const DEFAULT_TERMINAL_THEME_ID: TerminalThemeId = "one-dark";');
-    expect(source).toContain("themeId: DEFAULT_TERMINAL_THEME_ID");
-    expect(source).toContain('| "system"');
+  it("defaults new terminal settings to a dark terminal theme", () => {
+    expect(DEFAULT_TERMINAL_THEME_ID).toBe("one-dark");
   });
 });
